requirements.txt
lxml
numpy
opencv-python
opencv-contrib-python
pydantic-settings
pygeomag
pymap3d
svg.py
tqdm
utm
